// assets/gulp/src/types/IUniteConfiguration.ts
/**
 * Unite Configuration
 */
import { IUniteBuildConfiguration } from "./IUniteBuildConfiguration";
import { IUniteClientPackage } from "./IUniteClientPackage";
import { IUniteDirectories } from "./IUniteDirectories";

export interface IUniteConfiguration {
    uniteVersion: string;
    packageName: string;
    license: string;
    applicationFramework: string;
    moduleType: string;
    bundler: string;
    sourceLanguage: string;
    linter: string;
    unitTestRunner: string;
    unitTestFramework: string;
    unitTestEngine: string;
    e2eTestRunner: string;
    e2eTestFramework: string;
    cssPre: string;
    cssPost: string;
    server: string;
    taskManager: string;
    packageManager: string;
    ides: string[];

    sourceExtensions: string[];
    viewExtensions: string[];
    styleExtension: string;

    clientPackages: { [id: string]: IUniteClientPackage };

    dirs: IUniteDirectories;

    srcDistReplace: string;
    srcDistReplaceWith: string;

    buildConfigurations: { [id: string]: IUniteBuildConfiguration };
    platforms: { [id: string]: { [id: string]: any } };
}

// Generated by UniteJS